fix: changes according to requests in last PR.
