+++ constants/routes.js
@@ -2,22 +2,5 @@ import ICONS from "./icons";
 
 const ROUTES = {
   HOME: "/",
-  navbarPages: [
-    {
-      text: "Missions",
-      link: "/missions",
-      icon: ICONS.rocketicon,
-    },
-    {
-      text: "Leaderboard",
-      link: "/leaderboard",
-      icon: ICONS.trophyicon,
-    },
-    {
-      text: "Rewards",
-      link: "/rewards",
-      icon: ICONS.rewardsicon,
-    },
-  ],
 };
 export default ROUTES;
