+++ pages/layouts/AppLayout.js
@@ -4,11 +4,11 @@ import Header from "../../templates/Header/Header";
 
 export const AppLayout = ({ children }) => {
   return (
-    <>
+    <Grid>
       <Grid>
         <Header />
       </Grid>
       <Grid>{children}</Grid>
-    </>
+    </Grid>
   );
 };
